chore(opal): move Checkbox into Opal to fix Storybook process.env crash
The Table component imports Checkbox from @/refresh-components, which
imports @/lib/utils, which imports @/lib/constants — a file that
references process.env at the top level. This causes a "process is not
defined" crash in Storybook's browser environment when any story pulls
in the @opal/components barrel (which re-exports Table).

Moving Checkbox into @opal/components/checkbox severs this chain by
replacing the @/lib/utils import with @opal/utils. The old file at
@/refresh-components/inputs/Checkbox.tsx now re-exports from the Opal
source so all 16 existing consumers are unaffected.
