+++ web/lib/opal/src/components/table/TableQualifier.tsx
@@ -5,7 +5,7 @@ import { cn } from "@opal/utils";
 import { useTableSize } from "@opal/components/table/TableSizeContext";
 import type { IconFunctionComponent } from "@opal/types";
 import type { QualifierContentType } from "@opal/components/table/types";
-import Checkbox from "@/refresh-components/inputs/Checkbox";
+import Checkbox from "@opal/components/checkbox/components";
 
 interface TableQualifierProps {
   /** Content type displayed in the qualifier */
